Switch from pynput to libvinput (#134)
* Switch from pynput to libvinput

Signed-off-by: Slendi <slendi@socopon.com>

* Change modifier CLI args

* Fix: Update timing after adding key to word

Edge case where a non-allowed char key (e.g. number) typed fast enough to be considered a chord is added as the first character of a word, and timing is not initialized. When a non-chord key (e.g. modifier key) is then pressed (to end the word), the word is logged but there is no timing associated with it, which would cause a crash.

* Fix: device chords format

---------

Signed-off-by: Slendi <slendi@socopon.com>
Signed-off-by: Raymond Li <hi@raymond.li>
Co-authored-by: Raymond Li <hi@raymond.li>